requirements.txt
pyyaml
pandas
bluepy
subus2
google-api-python-client
google-auth-oauthlib